feat(netplay): wire RelayClient into NetplaySession (Slice 4)
Integrates the relay-assisted hole-punching state machine (Slice 3)
into the launcher's NetplaySession. Users can now start relay-assisted
host/join sessions that work through NAT without port forwarding.

Architecture:
  startRelayHost/Join() → .relay_connecting → stepRelay()
    → on success: transport.connectBound(peer_ip, peer_port, local_port)
    → .connecting (existing ENet handshake flow takes over)
    → .handshaking → .ping_exchanging → .waiting_confirmation → .launching

The relay path is a sibling to the direct-IP path — existing direct
host/join is completely unchanged. After relay handoff, the session
transitions to .connecting and reuses the existing ENet handshake
(version exchange, names, pings, config).

Changes:

enet_transport.zig:
  - Added connectBound() method — same as connect() but binds the ENet
    host to a specific local UDP port. Used after relay hole-punch to
    preserve the NAT mapping opened during hole-punching.
  - Option A handoff: relay client's UDP socket is closed (with
    SO_REUSEADDR), then ENet rebinds to the same port.

session.zig:
  - New SessionState: .relay_connecting
  - New fields: relay_client (?RelayClient), relay_list (?RelayList)
  - New methods:
    - startRelayHost(relay_source, port, training) — host via relay
    - startRelayJoin(relay_source, peer_identifier, training) — join via relay
    - getRoomCode() → ?[4]u8 (zzcaster host only)
    - isRelayMode() → bool
    - stepRelay() — drives RelayClient, handles ENet handoff on success
  - deinit() tears down relay sockets + list before transport
  - step() switch handles .relay_connecting

ui_waiting_for_peer.zig:
  - New helpers: startRelayHostSession(), startRelayJoinSession()
  - Waiting screen shows relay-specific info:
    - zzcaster host: room code + Copy button
    - cccaster host: IP:port + Copy button (same as direct)
    - relay client: 'Connecting via relay...'
    - hint: '(relay-assisted — no port forwarding needed)'
  - Direct-IP path unchanged

relay_client.zig:
  - Fixed FIONBIO constant type (u32 instead of c_long — value
    0x8004667E exceeds i32 max)
  - Fixed ioctlsocket extern signature (cmd: u32 instead of c_long)

Verified:
  zig build -Dtarget=x86-windows-gnu -Doptimize=ReleaseFast  # 12/12 OK
  zig build test                                              # 64/64 pass

Not yet done (Slice 5): UI mode toggle (Direct IP / Room Code),
room code input field, relay list loading from config.ini. The
relay path is programmatically accessible but no UI buttons trigger
it yet — that's Slice 5.

+++ src/net/relay_client.zig
@@ -39,7 +39,7 @@ const ws2_32 = struct {
     extern "ws2_32" fn sendto(s: c_int, buf: [*]const u8, len: c_int, flags: c_int, to: ?*const sockaddr_in, tolen: c_int) callconv(.winapi) c_int;
     extern "ws2_32" fn recvfrom(s: c_int, buf: [*]u8, len: c_int, flags: c_int, from: ?*sockaddr_in, fromlen: ?*c_int) callconv(.winapi) c_int;
     extern "ws2_32" fn select(nfds: c_int, readfds: ?*fd_set, writefds: ?*fd_set, exceptfds: ?*fd_set, timeout: ?*const timeval) callconv(.winapi) c_int;
-    extern "ws2_32" fn ioctlsocket(s: c_int, cmd: c_long, argp: *u32) callconv(.winapi) c_int;
+    extern "ws2_32" fn ioctlsocket(s: c_int, cmd: u32, argp: *u32) callconv(.winapi) c_int;
     extern "ws2_32" fn setsockopt(s: c_int, level: c_int, optname: c_int, optval: [*]const u8, optlen: c_int) callconv(.winapi) c_int;
     extern "ws2_32" fn getsockopt(s: c_int, level: c_int, optname: c_int, optval: [*]u8, optlen: *c_int) callconv(.winapi) c_int;
     extern "ws2_32" fn getsockname(s: c_int, name: ?*sockaddr_in, namelen: ?*c_int) callconv(.winapi) c_int;
@@ -72,7 +72,7 @@ const ws2_32 = struct {
     const SO_REUSEADDR: c_int = 0x0004;
     const SO_RCVTIMEO: c_int = 0x1006;
     const SO_ERROR: c_int = 0x1007;
-    const FIONBIO: c_long = 0x8004667E;
+    const FIONBIO: u32 = 0x8004667E;
 
     const WSAEWOULDBLOCK: c_int = 10035;
     const WSAEINPROGRESS: c_int = 10036;
